+++ examples/nextjs-app-router/CHANGELOG.md
@@ -1,5 +1,19 @@
 # example-nextjs-app-router
 
+## 0.0.7
+
+### Patch Changes
+
+- Updated dependencies
+  - @swr-login/core@0.7.0
+  - @swr-login/react@0.8.0
+  - @swr-login/adapter-cookie@0.2.5
+  - @swr-login/adapter-jwt@0.2.5
+  - @swr-login/plugin-oauth-github@0.2.5
+  - @swr-login/plugin-oauth-google@0.2.5
+  - @swr-login/plugin-passkey@0.2.5
+  - @swr-login/plugin-password@0.2.5
+
 ## 0.0.6
 
 ### Patch Changes